Adding competition

+++ src/main/java/com/itmo/repository/LevelRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository("levelRepository")
-public interface LevelRepository extends JpaRepository<Level, Integer> {
+public interface LevelRepository extends JpaRepository<Level, Long> {
     Level findByLevel(String level);
 
     @Override
